requirements.txt
fastapi
pydantic[email]
uvicorn[standard]
pandas
python-multipart
python-docx
Pillow
docxtpl
jinja2
hijri-converter
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.0
python-dateutil
python-dotenv
schedule
slowapi>=0.1.9

# PostgreSQL Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
